refactor: modularize answer service by splitting logic into specialized sub-services and interfaces
