+++ backend/src/modules/answer/interfaces/index.ts
@@ -1,1 +1,6 @@
-export * from './IAnswerService.js'
+export * from './IAnswerService.js';
+export * from './IAnswerReviewService.js';
+export * from './IAnswerApprovalService.js';
+export * from './IAnswerSubmissionService.js';
+export * from './IAnswerAiService.js';
+export * from './IAnswerFaqService.js';
