+++ backend/src/types.ts
@@ -24,6 +24,11 @@ const TYPES = {
   ModeratorQueueService: Symbol.for('ModeratorQueueService'),
   QuestionMaintenanceService: Symbol.for('QuestionMaintenanceService'),
   AnswerService: Symbol.for('AnswerService'),
+  AnswerReviewService: Symbol.for('AnswerReviewService'),
+  AnswerApprovalService: Symbol.for('AnswerApprovalService'),
+  AnswerSubmissionService: Symbol.for('AnswerSubmissionService'),
+  AnswerAiService: Symbol.for('AnswerAiService'),
+  AnswerFaqService: Symbol.for('AnswerFaqService'),
   ContextService: Symbol.for('ContextService'),
   CommentService: Symbol.for('CommentService'),
   RequestService: Symbol.for('RequestService'),
